requirements.txt
llama-index
llama-index-llms-ollama
llama-index-tools-mcp
ollama
nest_asyncio
